doing bladestorm

+++ src/main/java/xyz/destiall/mc/valorant/api/abilities/Smoke.java
@@ -8,5 +8,5 @@ public interface Smoke {
     void appear(Location location);
     void dissipate();
     Duration getSmokeDuration();
-    int getSmokeRange();
+    double getSmokeRange();
 }
